Split publish.ts into one module per phase
publish.ts had grown to 955 lines holding the orchestration, all six phases, a
per-section chunk writer, and the failure-cleanup path. Now 12 files, largest 168
lines, following the lib/<area>/ + index.ts barrel convention already used by
data-access's glossary/ and passage/.

  publish/
    index.ts           barrel
    start.ts           startPublish
    tick.ts            the tick loop, phase order, budget
    context.ts         PhaseContext / PhaseRunner
    validate-phase.ts
    snapshot-phase.ts
    artifact-phase.ts  passage chunking and the section cursor
    row-sections.ts    one chunk of a row-count-chunked section
    index-phase.ts
    manifest-phase.ts
    flip-phase.ts
    recover.ts         cleanup after a failed publish

Phases are now uniform — each takes a PhaseContext and returns the job as it
stands — so dispatch is a lookup keyed on phase rather than a switch that had to
know which phase wanted which arguments. `done` having no entry is what ends the
loop, so completion needs no special case.

The boundaries are the ones the pipeline already had. Each phase was independent
except through the job row, which is why they extracted without untangling
anything; the only genuinely coupled pair is the artifact phase and its row-section
writer, and those stayed adjacent.

Behaviour is unchanged, and verified rather than assumed: the 1ms-budget run still
takes exactly 10 ticks with the same cursor progression and file accumulation,
adoption still resumes an abandoned job at its own version, rebuild still restores
a corrupted snapshot to a matching fingerprint, and a validation hard-fail still
leaves versions, rows, and Storage objects untouched.
